com 7 to heroeschessgui
It's been quite a while since last commit, thus many changes and
improvements have been done. The only part remaining of move generation
and validity testing is preventing kings from attacking the opposite
king. Currently a king can capture the other but that player still
doesn't win the game.
Also, working on a revamp of the pawn promotion implementation, by
adding a promoteTo int to the Player class, which holds a value for the
type chosen as promotion. 0 if no promotion.
Added a menu, with exit and switches for displaying the three different
squares when moving/post move.
Implemented win game testing from a working check mate test method.