com 9 New version!
So many changes this has to be called a new version. Trying to get the
gameLoop woking for both human and network players, which means with and
without mouse listeners involved. This is a bit tricky.
Implemented four different backgrounds: grasshill, desert, winter and
swamp, that are selected by random.
Reorganized image loading, so exporting to runnable jar now works
perfectly.