com 10 v0.51
Tried to make one gameloop in Game that via lots of conditions (ifs)
works with both human and non-human players. So fra the result is pure
spaghetti, so I need to drastically improve it. Also, I feel there's far
too much gui-related calls inside the gameloop, that should be confined
to a gui class, like ChessGameGui or some new gui class/classes. Again,
my general feeling is that the code has started to lose a lot of the
object oriented feel, and started to be more procedural, which is not
the way I want to go.
As for the networking part, moves are being transferred between clients
and server, but the game crashes (null pointer ex) when attacking a
piece. I think it may have to do with the bitboards not being updated in
the right order, so when attack happens, the game returns an empty bit
thus a null pointer ex occurs. At least something like that.